Add files via upload
This a small portion of school management System using jdbc. There are two categories student and Teacher in this project. The student can login as well as register .After login the student can see the personal details and marks allotted to each subject.

The teacher can login as well as register .After login the teacher can see the personal details ,List of students and can enter  marks to the student according to their respective subject.